Linters, formatters, and a pre-commit hook for both sides
Rust: clippy gate at -D warnings. Mechanical warnings fixed via
cargo clippy --fix (59 files); pre-existing stylistic lints are
grandfathered in [workspace.lints.clippy] with burn-down instructions;
intentional dead code annotated with reasons. Also surfaced by the
sweep: a stray duplicate #[test] in skill_data.rs whose doc belonged
to loads_real_dist_files.

Frontend: Biome (format + lint in one binary, bun-native). Config
matches the existing style (2-space, double quotes, width 100,
Tailwind directives); bun run lint / bun run format scripts.

.githooks/pre-commit runs rustfmt + clippy when .rs files are staged
and biome --staged for web/dashboard files. Activate per clone with:
git config core.hooksPath .githooks (documented in README).

Verified: 1652 workspace tests pass, 46 frontend tests pass, tsc and
bun build clean, hook rejects unformatted code.

+++ crates/gameserver/src/model/mod.rs
@@ -1206,7 +1206,7 @@ impl Player {
             }
             b
         };
-        let acc_ev_step = hi_level_step(self.level as i32);
+        let acc_ev_step = hi_level_step(self.level);
         combat.accuracy = finalize(
             mods,
             Stat::AccuracyCombat,
